+++ anybox/recipe/openerp/tests/test_base.py
@@ -17,9 +17,9 @@ TEST_DIR = os.path.dirname(__file__)
 class TestingRecipe(BaseRecipe):
     """A subclass with just enough few defaults for unit testing."""
 
-    archive_filenames = {'6.1': 'blob-%s.tgz',
+    release_filenames = {'6.1': 'blob-%s.tgz',
                          '6.0': 'bl0b-%s.tgz'}
-    archive_nightly_filenames = {'6.1': '6-1-nightly-%s.tbz'}
+    nightly_filenames = {'6.1': '6-1-nightly-%s.tbz'}
 
 
 class TestBaseRecipe(RecipeTestCase):
@@ -54,13 +54,13 @@ class TestBaseRecipe(RecipeTestCase):
         recipe = self.recipe
         self.assertEquals(recipe.version_wanted, '6.1-1')
         self.assertDownloadUrl(
-            'http://nightly.openerp.com/6.1/releases/blob-6.1-1.tgz')
+            'http://nightly.odoo.com/6.1/releases/blob-6.1-1.tgz')
 
     def test_version_nightly_6_1(self):
         self.make_recipe(version='nightly 6.1 1234-5')
 
         self.assertDownloadUrl(
-            'http://nightly.openerp.com/6.1/nightly/src/'
+            'http://nightly.odoo.com/6.1/nightly/src/'
             '6-1-nightly-1234-5.tbz')
 
     def test_version_bzr_6_1(self):
@@ -186,7 +186,7 @@ class TestBaseRecipe(RecipeTestCase):
             version = conf.get('freeze', 'Babel')
         except NoOptionError:
             self.fail("Expected version of Babel egg not dumped !")
-        self.assertEquals(version, '0.123-dev')
+        self.assertTrue(version.startswith('0.123'))
 
     def test_freeze_egg_versions_merge(self):
         """Test that freezing of egg versions keeps eggs already dumped.
@@ -373,7 +373,7 @@ class TestBaseRecipe(RecipeTestCase):
                           "fakevcs+http://example.com/aeroolib@fakerev"
                           "#egg=aeroolib")
 
-    def test_finalize_addons_paths_odoo_layout(self):
+    def test_finalize_addons_paths_git_layout(self):
         self.make_recipe(
             version='git http://github.com/odoo/odoo.git odoo 7.0')
         self.recipe.version_detected = '7.0-somerev'
@@ -387,6 +387,9 @@ class TestBaseRecipe(RecipeTestCase):
         self.assertEquals(self.recipe.addons_paths,
                           [base_addons, odoo_addons, '/some/separate/addons'])
 
+        self.recipe._register_extra_paths()
+        self.assertEqual(self.recipe.extra_paths, [oerp_dir])
+
     def test_finalize_addons_paths_bzr_layout(self):
         self.make_recipe(
             version='bzr lp:openobject-server openerp last:1')
@@ -398,6 +401,25 @@ class TestBaseRecipe(RecipeTestCase):
         self.recipe.finalize_addons_paths(check_existence=False)
         self.assertEquals(self.recipe.addons_paths, [base_addons,
                                                      '/some/separate/addons'])
+
+    def test_finalize_addons_paths_order(self):
+        """Test finalize_addons_paths keeps addons_path order
+        Ensure we don't move odoo addons in addons_path if it has been
+        set as local
+        """
+        self.make_recipe(
+            version='git http://github.com/odoo/odoo.git odoo 7.0')
+        self.recipe.version_detected = '7.0-somerev'
+        oerp_dir = self.recipe.openerp_dir
+        base_addons = os.path.join(oerp_dir, 'openerp', 'addons')
+        odoo_addons = os.path.join(oerp_dir, 'addons')
+        os.makedirs(base_addons)
+        os.makedirs(odoo_addons)
+        self.recipe.addons_paths = ['/some/separate/addons',
+                                    odoo_addons]
+        self.recipe.finalize_addons_paths(check_existence=False)
+        self.assertEquals(self.recipe.addons_paths,
+                          [base_addons, '/some/separate/addons', odoo_addons])
 
     def test_finalize_addons_paths_60_layout(self):
         self.make_recipe(version='6.0.4')