+++ anybox/recipe/openerp/tests/test_utils.py
@@ -2,8 +2,9 @@ import unittest
 import tempfile
 import shutil
 import os
+from datetime import timedelta
 
-from ..utils import working_directory_keeper
+from ..utils import working_directory_keeper, total_seconds
 
 
 class WorkingDirectoryTestCase(unittest.TestCase):
@@ -23,3 +24,11 @@ class WorkingDirectoryTestCase(unittest.TestCase):
             self.assertNotEqual(os.getcwd(), current)
 
         self.assertEqual(os.getcwd(), current)
+
+
+class VariousTestCase(unittest.TestCase):
+
+    def test_total_seconds(self):
+        self.assertEqual(total_seconds(timedelta(1, 2)), 86402.0)
+        self.assertEqual(total_seconds(timedelta(0, -3)), -3.0)
+        self.assertEqual(total_seconds(timedelta(0, 12, 35000)), 12.035)